Refactor asset management in ServiceConfiguration
Removed the `ExtractAssetsOnce` method from `ServiceConfiguration.cs` and replaced it with a call to `EmbeddedResourceHelper.ExtractAssets` for centralized asset extraction. Updated logging messages and removed constants for asset filenames.

Introduced a new `ExtractAssets` method in `EmbeddedResourceHelper.cs` to handle embedded asset extraction with improved logging and validation. Consolidated individual asset extraction into a loop for better maintainability.

Updated asset validation logic to utilize a list of required assets, enhancing clarity and extensibility of the code.
